requirements/base.txt
fbs==0.8.4
PyQt5==5.9.2
Pillow>=6.1.0
pygame==1.9.6